Add login functionality and JWT module
Introduced a login controller and handler in the Authentication module to enable user login. This includes creating new Requests and Response DTOs to handle data transfer. Also added a JWT module to the system and updated the Local strategy in the authentication strategies. Made some minor changes to the package and tsconfig as part of the update.