Add 'verifiedAt' field to the User entity
This commit introduces a new 'verifiedAt' field to the User entity to keep track of when a user's account is verified. The addition of this field triggers a new event 'OnUserChangedVerifiedAtEvent' whenever the verifiedAt field is updated. It also includes a test case to ensure that 'verifiedAt' can be updated appropriately.