Add BaseTokenPayload type and TokenType enumeration
This commit introduces a new file, base-token-payload.type.ts, under the jwt domain. It comes with an enumeration for token types (TokenType) and a base interface for token payload (BaseTokenPayload). This interface includes type and data which is expected to serve as a standard for all types of jwt tokens in the application.