Refactor user assignment and remove unused token decorator
The user assignment has been refactored for better type casting in role-class-serializer.interceptor.ts and access-token.guard.ts. An unused token decorator was also removed in refresh.http.controller.ts. Finally, in the mock.sms.service.ts, a redundant parameter in the sendSMS and sendSMSs methods was replaced with an underscore to denote that it's not being used.