Refactor user mock creation to a dedicated test utility
This commit creates a new testing utility to generate user mocks, replacing the previous method which was directly included in the user entity file. The user mock creation method (`CreateUserMock`) has been moved from the user entity file to a new file in the `tests/utils` directory. The `CreateUserMock` function has been updated in all locations where it was previously called. Additionally, updates to the Jest configuration and `tsconfig.json` were made to support the changes.