Add script to setup template remote and update merge strategy
A bash script has been added which sets up a template remote for the project. This script checks and fetches the remote, creates a new branch from it, and merges it into the current branch. Furthermore, in the Github workflow file, the merge strategy during a template update has been changed to use 'recursive' with 'theirs' option to automatically resolve potential conflicts.